moved dlls stuff to MOApplication
